Feature: Add common exceptions

+++ src/main/java/fr/raph/musicstream/services/SongService.java
@@ -1,5 +1,7 @@
 package fr.raph.musicstream.services;
 
+import fr.raph.musicstream.exceptions.ObjectAlreadyExistException;
+import fr.raph.musicstream.exceptions.ObjectNotFoundException;
 import fr.raph.musicstream.mappers.SongMapper;
 import fr.raph.musicstream.models.Song;
 import fr.raph.musicstream.storage.SongRepository;
@@ -17,12 +19,15 @@ public class SongService {
     private final SongMapper SONG_MAPPER = SongMapper.SONG_MAPPER;
 
     public Song add(Song song) {
+        if (songRepository.findByName(song.getName()).isPresent()) { throw new ObjectAlreadyExistException("Song", song.getName()); }
+
         SongEntity songEntity = SONG_MAPPER.toSongEntity(song);
         SongEntity savedSongEntity = songRepository.save(songEntity);
         return SONG_MAPPER.toSong(savedSongEntity);
     }
 
     public void remove(String name) {
+        if (songRepository.findByName(name).isEmpty()) { throw new ObjectNotFoundException("Song", name); }
         songRepository.deleteByName(name);
     }
 
@@ -38,7 +43,7 @@ public class SongService {
     }
 
     public Song get(String name) {
-        SongEntity songEntity = songRepository.findByName(name);
+        SongEntity songEntity = songRepository.findByName(name).orElseThrow(() -> new ObjectNotFoundException("Song", name));
         return SONG_MAPPER.toSong(songEntity);
     }
 }
